fix: sitemap
- move sitemap into app folder
- /sitemap.xml now works
- add institution pages into sitemap
